python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bsvibe-auth"
version = "0.1.0"
description = "Unified authentication package for the BSVibe ecosystem — JWT verification based on Supabase Auth"
requires-python = ">=3.11"
dependencies = [
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.110.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["bsvibe_auth"]
